Fix Footer & Navigation routing issues
